Use an enumerated array to calculate IDs
This fixes an issue where two or more views of the same type would cause the duplicate(s) to not render. For example, if four text views were passed, only the first would render.

The zip function is being used instead of an array to allow for future adjustments, while ensuring that we can still iterate over it, since not all data structures start at zero. See here for more details: https://stackoverflow.com/questions/59295206/how-do-you-use-enumerated-with-foreach-in-swiftui